POAnimatedImage moved to 'ui-core'
